Extract logic in hook
